YES! External joysticks are handled now! <3 <3 <3

+++ ControllerApp/app/src/main/java/khr/easv/pokebotcontroller/app/gui/views/AbstractKnobView.java
@@ -147,9 +147,9 @@ public abstract class AbstractKnobView extends View {
     }
 
     /** Should return a difference in X position, relative to the center of the view */
-    protected abstract int getKnobX();
+    protected abstract float getKnobX();
     /** Should return a difference in Y position, relative to the center of the view */
-    protected abstract int getKnobY();
+    protected abstract float getKnobY();
 
     /** Call this to update the actual graphics */
     public void updateKnobPosition(){
